Add IBL support
### Purpose
Add support for ".exr" image extention for DomeLight

### Technical steps
1. Refactored:
    - moved `cache_image_file()` from `utils.cc` to `scene_delegate/image.cc`
    - moved `gf_matrix_from_transform()` from `utils.cc` to `scene_delegate/object.cc`
    - removed `utils.cc` as not needed.
2. Implemented more correct image caching.
3. Adjusted passing image for DomeLight.
4. Added copying hioOiio folder.

Pull Request https://projects.blender.org/BogdanNagirniak/blender/pulls/29

+++ source/blender/render/hydra/scene_delegate/blender_scene_delegate.cc
@@ -6,7 +6,6 @@
 #include "DEG_depsgraph_query.h"
 #include "DNA_scene_types.h"
 
-#include "..\utils.h"
 #include "blender_scene_delegate.h"
 
 namespace blender::render::hydra {
@@ -194,7 +193,7 @@ pxr::SdfPath BlenderSceneDelegate::prim_id(ID *id, const char *prefix) const
 {
   /* Making id of object in form like <prefix>_<pointer in 16 hex digits format> */
   char str[32];
-  snprintf(str, 32, "%s_%016llx", prefix, (uint64_t)id);
+  snprintf(str, 32, "%s_%016llx", prefix, (uintptr_t)id);
   return GetDelegateID().AppendElementString(str);
 }
